chore: Move telemetry python module
