+++ argilla-server/tests/unit/commons/test_telemetry.py
@@ -30,7 +30,7 @@ mock_request = Request(scope={"type": "http", "headers": {}})
 class TestSuiteTelemetry:
     async def test_create_client_with_server_id(self, mocker: MockerFixture):
         mock_server_id = uuid.uuid4()
-        mocker.patch("argilla_server.telemetry.get_server_id", return_value=mock_server_id)
+        mocker.patch("argilla_server.telemetry._client.get_server_id", return_value=mock_server_id)
 
         test_telemetry = TelemetryClient()
 
@@ -40,7 +40,7 @@ class TestSuiteTelemetry:
     async def test_track_data(self, mocker: MockerFixture):
         from argilla_server._version import __version__ as version
 
-        mock = mocker.patch("argilla_server.telemetry.send_telemetry")
+        mock = mocker.patch("argilla_server.telemetry._client.send_telemetry")
 
         telemetry = TelemetryClient()
 
@@ -54,7 +54,9 @@ class TestSuiteTelemetry:
         )
 
     async def test_track_api_request(self, test_telemetry: TelemetryClient, mocker: MockerFixture):
-        mocker.patch("argilla_server.telemetry.resolve_endpoint_path_for_request", return_value="/api/test/endpoint")
+        mocker.patch(
+            "argilla_server.telemetry._client.resolve_endpoint_path_for_request", return_value="/api/test/endpoint"
+        )
 
         request = Request(
             scope={
@@ -83,7 +85,9 @@ class TestSuiteTelemetry:
         )
 
     async def test_track_api_request_call_with_error(self, test_telemetry: TelemetryClient, mocker: MockerFixture):
-        mocker.patch("argilla_server.telemetry.resolve_endpoint_path_for_request", return_value="/api/test/endpoint")
+        mocker.patch(
+            "argilla_server.telemetry._client.resolve_endpoint_path_for_request", return_value="/api/test/endpoint"
+        )
 
         request = Request(
             scope={
@@ -110,7 +114,9 @@ class TestSuiteTelemetry:
     async def test_track_api_request_call_with_error_and_exception(
         self, test_telemetry: TelemetryClient, mocker: MockerFixture
     ):
-        mocker.patch("argilla_server.telemetry.resolve_endpoint_path_for_request", return_value="/api/test/endpoint")
+        mocker.patch(
+            "argilla_server.telemetry._client.resolve_endpoint_path_for_request", return_value="/api/test/endpoint"
+        )
 
         request = Request(
             scope={
